Remove use of proxy services to prepare change in API for QueryContext
