+++ webapp/sources/ldap-inventory/inventory-api/src/main/scala/com/normation/inventory/services/provisioning/CheckInventoryDigest.scala
@@ -161,46 +161,3 @@ trait GetKey {
     }
   }
 }
-
-class InventoryDigestServiceV1(
-    getExistingNode: NodeId => IOResult[Option[FullInventory]]
-) extends ParseInventoryDigestFileV1 with GetKey with CheckInventoryDigest {
-
-  /**
-   * Get key in V1 will get the key from inventory data.
-   * either an inventory has already been treated before, it will look into ldap repository
-   * or if there was no inventory before, it will look for the key in the received inventory
-   */
-  def getKey(receivedInventory: Inventory): IOResult[(SecurityToken, KeyStatus)] = {
-
-    def extractKey(node: NodeInventory): IOResult[SecurityToken] = {
-      for {
-        agent <- node.agents.headOption.notOptional("There is no public key in inventory")
-      } yield {
-        agent.securityToken
-      }
-    }
-
-    getExistingNode(receivedInventory.node.main.id).flatMap {
-      case Some(storedInventory) =>
-        val keyStatus = storedInventory.node.main.keyStatus
-        val inventory: NodeInventory = {
-          // if inventory was deleted, don't care, use new one
-          storedInventory.node.main.status match {
-            case RemovedInventory => receivedInventory.node
-            case _                => // in other case, check is the key update is valid
-              keyStatus match {
-                case UndefinedKey => // Trust On First Use (TOFU) (user may have reseted, etc)
-                  receivedInventory.node
-                // Certified node always use stored inventory key
-                case CertifiedKey => storedInventory.node
-              }
-          }
-        }
-        extractKey(inventory).map((_, keyStatus))
-      case _                     =>
-        val status = receivedInventory.node.main.keyStatus
-        extractKey(receivedInventory.node).map((_, status))
-    }
-  }
-}
