+++ webapp/sources/rudder/rudder-rest/src/main/scala/com/normation/rudder/web/services/CurrentUser.scala
@@ -90,4 +90,5 @@ object CurrentUser extends SessionVar[Option[RudderUserDetail]]({
       case Some(u) => u.apiAuthz
     }
   }
+
 }
