+++ webapp/sources/rudder/rudder-web/src/main/scala/bootstrap/liftweb/RudderConfig.scala
@@ -111,7 +111,6 @@ import com.normation.rudder.facts.nodes.GitNodeFactStorageImpl
 import com.normation.rudder.facts.nodes.HistorizeNodeState
 import com.normation.rudder.facts.nodes.LdapNodeFactStorage
 import com.normation.rudder.facts.nodes.NodeFactChangeEventCallback
-import com.normation.rudder.facts.nodes.NodeFactFullInventoryRepositoryProxy
 import com.normation.rudder.facts.nodes.NodeFactInventorySaver
 import com.normation.rudder.facts.nodes.NodeFactRepository
 import com.normation.rudder.facts.nodes.NodeInfoServiceProxy
@@ -1151,7 +1150,6 @@ object RudderConfig extends Loggable {
   val eventLogDetailsService:              EventLogDetailsService                     = rci.eventLogDetailsService
   val eventLogRepository:                  EventLogRepository                         = rci.eventLogRepository
   val findExpectedReportRepository:        FindExpectedReportRepository               = rci.findExpectedReportRepository
-  val fullInventoryRepository:             FullInventoryRepository[Unit]              = rci.fullInventoryRepository
   val gitRevisionProvider:                 GitRevisionProvider                        = rci.gitRevisionProvider
   val healthcheckNotificationService:      HealthcheckNotificationService             = rci.healthcheckNotificationService
   val historizeNodeCountBatch:             IOResult[Unit]                             = rci.historizeNodeCountBatch
@@ -1314,7 +1312,6 @@ case class RudderServiceApi(
     personIdentService:                  PersonIdentService,
     gitRevisionProvider:                 GitRevisionProvider,
     logDisplayer:                        LogDisplayer,
-    fullInventoryRepository:             FullInventoryRepository[Unit],
     acceptedNodeQueryProcessor:          QueryProcessor,
     categoryHierarchyDisplayer:          CategoryHierarchyDisplayer,
     dynGroupService:                     DynGroupService,
@@ -1398,7 +1395,7 @@ object RudderConfigInit {
   /**
    * Catch all exception during initialization that would prevent initialization.
    * All exception are catched and will stop the application on boot.
-   * 
+   *
    * Throwing this is more transparent, otherwise the raw error could be unclear
    */
   sealed abstract class InitError(val msg: String, val cause: Option[Throwable])
@@ -1720,11 +1717,9 @@ object RudderConfigInit {
     lazy val nodeApiService = new NodeApiService(
       rwLdap,
       nodeFactRepository,
-      factFullInventoryRepo,
       roNodeGroupRepository,
       roLDAPParameterRepository,
       roAgentRunsRepository,
-      woFactNodeRepository,
       ldapEntityMapper,
       stringUuidGenerator,
       nodeDit,
@@ -1738,8 +1733,6 @@ object RudderConfigInit {
       reportingServiceImpl,
       queryProcessor,
       inventoryQueryChecker,
-      asyncDeploymentAgent,
-      userService,
       () => configService.rudder_global_policy_mode().toBox,
       RUDDER_RELAY_API
     )
@@ -1966,7 +1959,8 @@ object RudderConfigInit {
         pipelinedInventoryParser,
         inventorySaver,
         maxParallel,
-        new InventoryDigestServiceV1((id: NodeId) => factFullInventoryRepo.get(id)),
+        // it's always rudder doing these checking queries
+        new InventoryDigestServiceV1((id: NodeId) => nodeFactRepository.get(id)),
         checkLdapAlive
       )
     }
@@ -2005,8 +1999,6 @@ object RudderConfigInit {
         KEEP_DELETED_NODE_FACT_DURATION
       )
     }
-
-    lazy val factFullInventoryRepo = new NodeFactFullInventoryRepositoryProxy(nodeFactRepository)
 
     lazy val archiveApi = {
       val archiveBuilderService =
@@ -2098,6 +2090,7 @@ object RudderConfigInit {
           restDataSerializer,
           nodeApiService,
           nodeInheritedProperties,
+          uuidGen,
           DeleteMode.Erase // only supported mode for Rudder 8.0
         ),
         new ParameterApi(restExtractorService, zioJsonExtractor, parameterApiService2, parameterApiService14),
@@ -2534,7 +2527,7 @@ object RudderConfigInit {
       )
     }
 
-    lazy val nodeGridImpl = new NodeGrid(factFullInventoryRepo, nodeFactInfoService, configService)
+    lazy val nodeGridImpl = new NodeGrid(nodeFactRepository, nodeFactInfoService, configService)
 
     lazy val modificationService      =
       new ModificationService(logRepository, gitModificationRepository, itemArchiveManagerImpl, uuidGen)
@@ -2892,7 +2885,6 @@ object RudderConfigInit {
         ruleApplicationStatusImpl,
         roParameterServiceImpl,
         interpolationCompiler,
-        factFullInventoryRepo,
         globalComplianceModeService,
         globalAgentRunService,
         reportingServiceImpl,
@@ -3570,7 +3562,6 @@ object RudderConfigInit {
       personIdentServiceImpl,
       gitRevisionProviderImpl,
       logDisplayerImpl,
-      factFullInventoryRepo,
       queryProcessor,
       categoryHierarchyDisplayerImpl,
       dynGroupServiceImpl,
